move test functions and test utils functions in their own package
